major change to header structure. The inclusion of headers is not at all fixed. A bunch of edits have been made to function names in types, I guess this will take a bit longer to get back together :)
